Fix genDate generator frequency
I discovered that the generator was generating random years but the day
was always December 31st, January 31st, December 1st or January 1st (in
frequency order).
This was because of
https://github.com/purescript/purescript-gen/issues/22 which is now
fixed.
However, reading at the current implementation, there is still a slight
skew toward first days of months following a month with only 30 days or
less. This is because `genYear <*> genMonth <*> genDay` could generate
something like 2020-04-31 which then get converted to 2020-05-01 by
`canonicalDate`. This means that 2020-05-01 has twice as much chance to
get generated as other dates.

The fix is to pick a day number between 1 and 365 (or 366 for leap
years) instead.